// src/pages/demo/typings.d.ts
export type tabKeyType = 'articles' | 'applications' | 'projects';

export interface TagType {
  key: string;
  label: string;
}

export type GeographicType = {
  province: {
    label: string;
    key: string;
  };
  city: {
    label: string;
    key: string;
  };
};

export type NoticeType = {
  id: string;
  title: string;
  logo: string;
  description: string;
  updatedAt: string;
  member: string;
  href: string;
  memberLink: string;
};

export type CurrentUser = {
  name: string;
  userid: string;
  notice: NoticeType[];
  email: string;
  signature: string;
  title: string;
  group: string;
  tags: TagType[];
  notifyCount: number;
  unreadCount: number;
  country: string;
  geographic: GeographicType;
  address: string;
  phone: string;
};

export type Member = {
  name: string;
  id: string;
};

export type ListItemDataType = {
  id: string;
  title: string;
  status: 'normal' | 'exception' | 'active' | 'success';
  percent: number;
  members: Member[];
};